wallet: (move-only) extract out migration specific CWallet methods
From wallet/wallet.cpp to wallet/migration.cpp.

This patch should be reviewed with --color-moved=dimmed-zebra option.
