+++ src/wallet/wallet.cpp
@@ -3517,16 +3517,6 @@ std::vector<WalletDescriptor> CWallet::GetWalletDescriptors(const CScript& scrip
     return descs;
 }
 
-LegacyDataSPKM* CWallet::GetLegacyDataSPKM() const
-{
-    if (IsWalletFlagSet(WALLET_FLAG_DESCRIPTORS)) {
-        return nullptr;
-    }
-    auto it = m_internal_spk_managers.find(OutputType::LEGACY);
-    if (it == m_internal_spk_managers.end()) return nullptr;
-    return dynamic_cast<LegacyDataSPKM*>(it->second);
-}
-
 void CWallet::AddScriptPubKeyMan(const uint256& id, std::unique_ptr<ScriptPubKeyMan> spkm_man)
 {
     // Add spkm_man to m_spk_managers before calling any method
@@ -3841,311 +3831,6 @@ util::Result<std::reference_wrapper<DescriptorScriptPubKeyMan>> CWallet::AddWall
     MarkDirty();
 
     return std::reference_wrapper(*spk_man);
-}
-
-bool CWallet::MigrateToSQLite(bilingual_str& error)
-{
-    AssertLockHeld(cs_wallet);
-
-    WalletLogPrintf("Migrating wallet storage database from BerkeleyDB to SQLite.\n");
-
-    if (m_database->Format() == "sqlite") {
-        error = _("Error: This wallet already uses SQLite");
-        return false;
-    }
-
-    // Get all of the records for DB type migration
-    std::unique_ptr<DatabaseBatch> batch = m_database->MakeBatch();
-    std::unique_ptr<DatabaseCursor> cursor = batch->GetNewCursor();
-    std::vector<std::pair<SerializeData, SerializeData>> records;
-    if (!cursor) {
-        error = _("Error: Unable to begin reading all records in the database");
-        return false;
-    }
-    DatabaseCursor::Status status = DatabaseCursor::Status::FAIL;
-    while (true) {
-        DataStream ss_key{};
-        DataStream ss_value{};
-        status = cursor->Next(ss_key, ss_value);
-        if (status != DatabaseCursor::Status::MORE) {
-            break;
-        }
-        SerializeData key(ss_key.begin(), ss_key.end());
-        SerializeData value(ss_value.begin(), ss_value.end());
-        records.emplace_back(key, value);
-    }
-    cursor.reset();
-    batch.reset();
-    if (status != DatabaseCursor::Status::DONE) {
-        error = _("Error: Unable to read all records in the database");
-        return false;
-    }
-
-    // Close this database and delete the file
-    fs::path db_path = fs::PathFromString(m_database->Filename());
-    m_database->Close();
-    fs::remove(db_path);
-
-    // Generate the path for the location of the migrated wallet
-    // Wallets that are plain files rather than wallet directories will be migrated to be wallet directories.
-    const fs::path wallet_path = fsbridge::AbsPathJoin(GetWalletDir(), fs::PathFromString(m_name));
-
-    // Make new DB
-    DatabaseOptions opts;
-    opts.require_create = true;
-    opts.require_format = DatabaseFormat::SQLITE;
-    DatabaseStatus db_status;
-    std::unique_ptr<WalletDatabase> new_db = MakeDatabase(wallet_path, opts, db_status, error);
-    assert(new_db); // This is to prevent doing anything further with this wallet. The original file was deleted, but a backup exists.
-    m_database.reset();
-    m_database = std::move(new_db);
-
-    // Write existing records into the new DB
-    batch = m_database->MakeBatch();
-    bool began = batch->TxnBegin();
-    assert(began); // This is a critical error, the new db could not be written to. The original db exists as a backup, but we should not continue execution.
-    for (const auto& [key, value] : records) {
-        if (!batch->Write(std::span{key}, std::span{value})) {
-            batch->TxnAbort();
-            m_database->Close();
-            fs::remove(m_database->Filename());
-            assert(false); // This is a critical error, the new db could not be written to. The original db exists as a backup, but we should not continue execution.
-        }
-    }
-    bool committed = batch->TxnCommit();
-    assert(committed); // This is a critical error, the new db could not be written to. The original db exists as a backup, but we should not continue execution.
-    return true;
-}
-
-std::optional<MigrationData> CWallet::GetDescriptorsForLegacy(bilingual_str& error) const
-{
-    AssertLockHeld(cs_wallet);
-
-    LegacyDataSPKM* legacy_spkm = GetLegacyDataSPKM();
-    if (!Assume(legacy_spkm)) {
-        // This shouldn't happen
-        error = Untranslated(STR_INTERNAL_BUG("Error: Legacy wallet data missing"));
-        return std::nullopt;
-    }
-
-    std::optional<MigrationData> res = legacy_spkm->MigrateToDescriptor();
-    if (res == std::nullopt) {
-        error = _("Error: Unable to produce descriptors for this legacy wallet. Make sure to provide the wallet's passphrase if it is encrypted.");
-        return std::nullopt;
-    }
-    return res;
-}
-
-util::Result<void> CWallet::ApplyMigrationData(WalletBatch& local_wallet_batch, MigrationData& data)
-{
-    AssertLockHeld(cs_wallet);
-
-    LegacyDataSPKM* legacy_spkm = GetLegacyDataSPKM();
-    if (!Assume(legacy_spkm)) {
-        // This shouldn't happen
-        return util::Error{Untranslated(STR_INTERNAL_BUG("Error: Legacy wallet data missing"))};
-    }
-
-    // Note: when the legacy wallet has no spendable scripts, it must be empty at the end of the process.
-    bool has_spendable_material = !data.desc_spkms.empty() || data.master_key.key.IsValid();
-
-    // Get all invalid or non-watched scripts that will not be migrated
-    std::set<CTxDestination> not_migrated_dests;
-    for (const auto& script : legacy_spkm->GetNotMineScriptPubKeys()) {
-        CTxDestination dest;
-        if (ExtractDestination(script, dest)) not_migrated_dests.emplace(dest);
-    }
-
-    // When the legacy wallet has no spendable scripts, the main wallet will be empty, leaving its script cache empty as well.
-    // The watch-only and/or solvable wallet(s) will contain the scripts in their respective caches.
-    if (!data.desc_spkms.empty()) Assume(!m_cached_spks.empty());
-    if (!data.watch_descs.empty()) Assume(!data.watchonly_wallet->m_cached_spks.empty());
-    if (!data.solvable_descs.empty()) Assume(!data.solvable_wallet->m_cached_spks.empty());
-
-    for (auto& desc_spkm : data.desc_spkms) {
-        if (m_spk_managers.contains(desc_spkm->GetID())) {
-            return util::Error{_("Error: Duplicate descriptors created during migration. Your wallet may be corrupted.")};
-        }
-        uint256 id = desc_spkm->GetID();
-        AddScriptPubKeyMan(id, std::move(desc_spkm));
-    }
-
-    // Remove the LegacyScriptPubKeyMan from disk
-    if (!legacy_spkm->DeleteRecordsWithDB(local_wallet_batch)) {
-        return util::Error{_("Error: cannot remove legacy wallet records")};
-    }
-
-    // Remove the LegacyScriptPubKeyMan from memory
-    m_spk_managers.erase(legacy_spkm->GetID());
-    m_external_spk_managers.clear();
-    m_internal_spk_managers.clear();
-
-    // Setup new descriptors (only if we are migrating any key material)
-    SetWalletFlagWithDB(local_wallet_batch, WALLET_FLAG_DESCRIPTORS | WALLET_FLAG_LAST_HARDENED_XPUB_CACHED);
-    if (has_spendable_material && !IsWalletFlagSet(WALLET_FLAG_DISABLE_PRIVATE_KEYS)) {
-        // Use the existing master key if we have it
-        if (data.master_key.key.IsValid()) {
-            SetupDescriptorScriptPubKeyMans(local_wallet_batch, data.master_key);
-        } else {
-            // Setup with a new seed if we don't.
-            SetupOwnDescriptorScriptPubKeyMans(local_wallet_batch);
-        }
-    }
-
-    // Get best block locator so that we can copy it to the watchonly and solvables
-    // Note: The best block locator was introduced in #152 so ancient wallets do not have it
-    CBlockLocator best_block_locator;
-    (void)local_wallet_batch.ReadBestBlock(best_block_locator);
-
-    // Update m_txos to match the descriptors remaining in this wallet
-    m_txos.clear();
-    RefreshAllTXOs();
-
-    // Check if the transactions in the wallet are still ours. Either they belong here, or they belong in the watchonly wallet.
-    // We need to go through these in the tx insertion order so that lookups to spends works.
-    std::vector<Txid> txids_to_delete;
-    std::unique_ptr<WalletBatch> watchonly_batch;
-    if (data.watchonly_wallet) {
-        watchonly_batch = std::make_unique<WalletBatch>(data.watchonly_wallet->GetDatabase());
-        if (!watchonly_batch->TxnBegin()) return util::Error{strprintf(_("Error: database transaction cannot be executed for wallet %s"), data.watchonly_wallet->GetName())};
-        // Copy the next tx order pos to the watchonly wallet
-        LOCK(data.watchonly_wallet->cs_wallet);
-        data.watchonly_wallet->nOrderPosNext = nOrderPosNext;
-        watchonly_batch->WriteOrderPosNext(data.watchonly_wallet->nOrderPosNext);
-        // Write the locator record. An empty locator is valid and triggers rescan on load.
-        if (!watchonly_batch->WriteBestBlock(best_block_locator)) {
-            return util::Error{_("Error: Unable to write watchonly wallet best block locator record")};
-        }
-    }
-    std::unique_ptr<WalletBatch> solvables_batch;
-    if (data.solvable_wallet) {
-        solvables_batch = std::make_unique<WalletBatch>(data.solvable_wallet->GetDatabase());
-        if (!solvables_batch->TxnBegin()) return util::Error{strprintf(_("Error: database transaction cannot be executed for wallet %s"), data.solvable_wallet->GetName())};
-        // Write the locator record. An empty locator is valid and triggers rescan on load.
-        if (!solvables_batch->WriteBestBlock(best_block_locator)) {
-            return util::Error{_("Error: Unable to write solvable wallet best block locator record")};
-        }
-    }
-    for (const auto& [_pos, wtx] : wtxOrdered) {
-        // Check it is the watchonly wallet's
-        // solvable_wallet doesn't need to be checked because transactions for those scripts weren't being watched for
-        bool is_mine = IsMine(*wtx->tx) || IsFromMe(*wtx->tx);
-        if (data.watchonly_wallet) {
-            LOCK(data.watchonly_wallet->cs_wallet);
-            if (data.watchonly_wallet->IsMine(*wtx->tx) || data.watchonly_wallet->IsFromMe(*wtx->tx)) {
-                // Add to watchonly wallet
-                const Txid& hash = wtx->GetHash();
-                const CWalletTx& to_copy_wtx = *wtx;
-                if (!data.watchonly_wallet->LoadToWallet(hash, [&](CWalletTx& ins_wtx, bool new_tx) EXCLUSIVE_LOCKS_REQUIRED(data.watchonly_wallet->cs_wallet) {
-                    if (!new_tx) return false;
-                    ins_wtx.SetTx(to_copy_wtx.tx);
-                    ins_wtx.CopyFrom(to_copy_wtx);
-                    return true;
-                })) {
-                    return util::Error{strprintf(_("Error: Could not add watchonly tx %s to watchonly wallet"), wtx->GetHash().GetHex())};
-                }
-                watchonly_batch->WriteTx(data.watchonly_wallet->mapWallet.at(hash));
-                // Mark as to remove from the migrated wallet only if it does not also belong to it
-                if (!is_mine) {
-                    txids_to_delete.push_back(hash);
-                    continue;
-                }
-            }
-        }
-        if (!is_mine) {
-            // Both not ours and not in the watchonly wallet
-            return util::Error{strprintf(_("Error: Transaction %s in wallet cannot be identified to belong to migrated wallets"), wtx->GetHash().GetHex())};
-        }
-        // Rewrite the transaction so that anything that may have changed about it in memory also persists to disk
-        local_wallet_batch.WriteTx(*wtx);
-    }
-
-    // Do the removes
-    if (txids_to_delete.size() > 0) {
-        if (auto res = RemoveTxs(local_wallet_batch, txids_to_delete); !res) {
-            return util::Error{_("Error: Could not delete watchonly transactions. ") + util::ErrorString(res)};
-        }
-    }
-
-    // Pair external wallets with their corresponding db handler
-    std::vector<std::pair<std::shared_ptr<CWallet>, std::unique_ptr<WalletBatch>>> wallets_vec;
-    if (data.watchonly_wallet) wallets_vec.emplace_back(data.watchonly_wallet, std::move(watchonly_batch));
-    if (data.solvable_wallet) wallets_vec.emplace_back(data.solvable_wallet, std::move(solvables_batch));
-
-    // Write address book entry to disk
-    auto func_store_addr = [](WalletBatch& batch, const CTxDestination& dest, const CAddressBookData& entry) {
-        auto address{EncodeDestination(dest)};
-        if (entry.purpose) batch.WritePurpose(address, PurposeToString(*entry.purpose));
-        if (entry.label) batch.WriteName(address, *entry.label);
-        for (const auto& [id, request] : entry.receive_requests) {
-            batch.WriteAddressReceiveRequest(dest, id, request);
-        }
-        if (entry.previously_spent) batch.WriteAddressPreviouslySpent(dest, true);
-    };
-
-    // Check the address book data in the same way we did for transactions
-    std::vector<CTxDestination> dests_to_delete;
-    for (const auto& [dest, record] : m_address_book) {
-        // Ensure "receive" entries that are no longer part of the original wallet are transferred to another wallet
-        // Entries for everything else ("send") will be cloned to all wallets.
-        bool require_transfer = record.purpose == AddressPurpose::RECEIVE && !IsMine(dest);
-        bool copied = false;
-        for (auto& [wallet, batch] : wallets_vec) {
-            LOCK(wallet->cs_wallet);
-            if (require_transfer && !wallet->IsMine(dest)) continue;
-
-            // Copy the entire address book entry
-            wallet->m_address_book[dest] = record;
-            func_store_addr(*batch, dest, record);
-
-            copied = true;
-            // Only delete 'receive' records that are no longer part of the original wallet
-            if (require_transfer) {
-                dests_to_delete.push_back(dest);
-                break;
-            }
-        }
-
-        // Fail immediately if we ever found an entry that was ours and cannot be transferred
-        // to any of the created wallets (watch-only, solvable).
-        // Means that no inferred descriptor maps to the stored entry. Which mustn't happen.
-        if (require_transfer && !copied) {
-
-            // Skip invalid/non-watched scripts that will not be migrated
-            if (not_migrated_dests.contains(dest)) {
-                dests_to_delete.push_back(dest);
-                continue;
-            }
-
-            return util::Error{_("Error: Address book data in wallet cannot be identified to belong to migrated wallets")};
-        }
-    }
-
-    // Persist external wallets address book entries
-    for (auto& [wallet, batch] : wallets_vec) {
-        if (!batch->TxnCommit()) {
-            return util::Error{strprintf(_("Error: Unable to write data to disk for wallet %s"), wallet->GetName())};
-        }
-    }
-
-    // Remove the things to delete in this wallet
-    if (dests_to_delete.size() > 0) {
-        for (const auto& dest : dests_to_delete) {
-            if (!DelAddressBookWithDB(local_wallet_batch, dest)) {
-                return util::Error{_("Error: Unable to remove watchonly address book data")};
-            }
-        }
-    }
-
-    // If there was no key material in the main wallet, there should be no records on it anymore.
-    // This wallet will be discarded at the end of the process. Only wallets that contain the
-    // migrated records will be presented to the user.
-    if (!has_spendable_material) {
-        if (!m_address_book.empty()) return util::Error{_("Error: Not all address book records were migrated")};
-        if (!mapWallet.empty()) return util::Error{_("Error: Not all transaction records were migrated")};
-    }
-
-    return {}; // all good
 }
 
 bool CWallet::CanGrindR() const
